+++ packages/terra-site/src/examples/content-container/ContentContainerStandard.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import ContentContainer from 'terra-content-container';
-import Button from 'terra-button';
+import Heading from 'terra-heading';
 
 const display1 = <p key="1">display1display1display1display1display1display1display1display1display1display1display1display1display1display1</p>;
 const display2 = <p key="2">display2display2display2display2display2display2display2display2display2display2display2display2display2display2</p>;
@@ -14,10 +14,10 @@ const displays1 = [display1, display2, display3, display4, display5, display6];
 const displays2 = displays1;
 const displays3 = displays1;
 
-const button1 = <Button text="header button" />;
+const contentheader = <Heading level={1} size={'large'}>Display Header</Heading>;
 
 const container = () => (
-  <ContentContainer header={button1}>
+  <ContentContainer header={contentheader}>
     <div key="1">
       {displays1}
     </div>